use long instead of in for array size

+++ bubblesort.c
@@ -4,9 +4,9 @@
 /*                           tri a bulle séquentiel                           */
 /******************************************************************************/
 
-void bubbleSort(int *arr, int n, int (*compare)(int, int)) {
-    for (int i = n - 1; i >= 0; --i) {
-        for (int j = 0; j < i; ++j) {
+void bubbleSort(int *arr, long n, int (*compare)(int, int)) {
+    for (long i = n - 1; i >= 0; --i) {
+        for (long j = 0; j < i; ++j) {
             if (compare(arr[j], arr[j + 1])) {
                 int tmp = arr[j];
                 arr[j] = arr[j + 1];
